+++ csc/2016/vsv/src/handlers/TaskResultToResponseBuilder.kt
@@ -1,4 +1,4 @@
-package handlers.requests
+package handlers
 
 import communication.CommunicationProtos
